Gantt Chart Component
Creating gantt chart component
